+++ R/mirai-package.R
@@ -34,8 +34,8 @@
 #' @author Charlie Gao \email{charlie.gao@@shikokuchuo.net}
 #'     (\href{https://orcid.org/0000-0002-0750-061X}{ORCID})
 #'
-#' @importFrom nanonext call_aio context is_nul_byte msleep random recv request
-#'     send send_aio socket stop_aio unresolved .mirai_scm
+#' @importFrom nanonext call_aio context is_error_value msleep random recv
+#'     request send send_aio socket stop_aio unresolved .mirai_scm
 #'
 #' @docType package
 #' @name mirai-package
@@ -54,7 +54,7 @@ NULL
 }
 
 #' @export
-nanonext::is_nul_byte
+nanonext::is_error_value
 
 #' @export
 nanonext::unresolved